@@ -7,7 +7,7 @@
 	"author": "Pablo Klaschka",
 	"scripts": {
 		"build": "npm run compile && xdpm package dist",
-		"compile": "webpack --config webpack.config.js --mode production",
+		"compile": "cross-env SET NODE_OPTIONS=--openssl-legacy-provider && webpack --config webpack.config.js --mode production",
 		"debug:install": "webpack --config webpack.config.js --mode production && xdpm install dist -o",
 		"debug:watch": "xdpm watch dist | webpack --config webpack.config.js --mode production --watch",
 		"debug:watch-prerelease": "xdpm watch dist -w p | webpack --config webpack.config.js --mode production --watch",
@@ -64,6 +64,7 @@
 		"babel-loader": "^8.2.3",
 		"babel-plugin-transform-react-jsx": "^6.24.1",
 		"chalk": "^4.1.2",
+		"cross-env": "^7.0.3",
 		"css-loader": "^5.2.7",
 		"cypress": "^9.2.0",
 		"dita-ot-helper": "^0.13.0",
@@ -81,14 +82,14 @@
 		"prettier-plugin-package": "^1.3.0",
 		"prettier-plugin-style-order": "^0.2.2",
 		"sass": "^1.45.1",
-		"sass-loader": "^10.1.1",
+		"sass-loader": "^12.4.0",
 		"shelljs": "^0.8.4",
 		"sloc": "^0.2.1",
 		"source-map-loader": "^3.0.0",
-		"style-loader": "^2.0.0",
+		"style-loader": "^3.3.1",
 		"ts-loader": "^8.3.0",
 		"typescript": "^4.5.4",
-		"webpack": "^4.46.0",
+		"webpack": "^5.65.0",
 		"webpack-cli": "^4.9.1",
 		"webpack-dev-server": "^4.7.1"
 	}
